refactor: split edge case tests
